Add GEC-only annotations (WIP)
Naively removing fluency edits may result in ungrammatical sentence. In
order to avoid that, annotators validated the whole corpus with the
fluency edits removed. They made edits to ensure that sentences stay
grammatical even after fluency is removed.

This branch contains the results of their work.

The commit temporarily adds two folders.

`results-gec-only.src` contains snippets that were handed to annotators
for the gec-only annotation. These snippets are:

* should contain all snippets added in v1 and v2
* fluency annotations were *removed*
* have no detailed annotations
* have to fixes applied to v1 and v2 since the data collection

`results-gec-only.tgt` is the result of the annotators' work:

* should fix errors introduced by removing fluency annotations
* sometimes annotators make edits unrelated to fluency removal (seems like they were fixing random errors along the way)

These are subtasks for integrating GEC-only into the release:

[ ] Copy detailed annotations here
[ ] Integrate fixes made to the main corpus here
[ ] Backport some edits from GEC-only to the main corpus (where edits are bugfixes and are not related to fluency removal)
[ ] Check for missing snippets and fix it
[ ] Update docs and find a proper place for GEC-only